GH-56: separate command configuration object
