+++ src/__tests__/GameBoardBuilder.test.ts
@@ -1,5 +1,5 @@
 import GameBoardBuilder from '@bot/gameboard/GameBoardBuilder';
-import localize from '@config/localize';
+import localize from '@i18n/localize';
 import AI from '@tictactoe/AI';
 import { Player } from '@tictactoe/Player';
