+++ src/config/ConfigProvider.ts
@@ -1,6 +1,6 @@
-import path from 'path';
-import fs from 'fs';
 import Config from '@config/Config';
+import fs from 'fs';
+import path from 'path';
 
 /**
  * Manages the configuration loading to operate the bot.
